% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarBalDiagEval.R
\docType{package}
\name{CovarBalDiagEval-package}
\alias{CovarBalDiagEval}
\alias{CovarBalDiagEval-package}
\title{CovarBalDiagEval: Evaluating the fitness of various statistics to measure covariate balance.}
\description{
A framework for evaluating the use of different statistics (starting with standardized mean difference [SMD]) to measure covariate balance.
}
\keyword{internal}
