% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  databaseId,
  packageName,
  randomSeed = 123,
  databaseName = databaseId,
  databaseDescription = databaseId,
  maxCores = parallel::detectCores() - 1,
  maxCohortSize = 1e+05,
  createCohorts = FALSE,
  synthesizePositiveControls = FALSE,
  createCohortMethodObjects = FALSE,
  generateAnalysisObjects = FALSE,
  synthesizeAndExportResults = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where outcome data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the \code{cohortDatabaseSchema}.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{packageName}{The name of the package, added for flexibility in package naming convention with find/replace.}

\item{randomSeed}{The seed to use for random samples - for greater reproducibility.}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{maxCohortSize}{The largest cohort size to be used when creating \code{CohortMethod} data objects.}

\item{createCohorts}{Create the exposure and outcome cohorts?}

\item{synthesizePositiveControls}{Create synthetic positive controls using \code{MethodEvaluation} package?}

\item{createCohortMethodObjects}{Create the CohortMethod data objects?}

\item{generateAnalysisObjects}{Computes the balance objects, fits outcome models, and generates population meta-data.}

\item{synthesizeAndExportResults}{Synthesize results across analyses?}

\item{verbose}{Should verbose logging be used?}
}
\description{
Execute the study
}
\details{
This function executes the study.
}
