% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize.R
\name{synthesizeResults}
\alias{synthesizeResults}
\title{Combines results across databases.}
\usage{
synthesizeResults(
  baseOutputFolder,
  dbIds,
  packageName,
  finalOutputFolder = getResultsFolderPath(baseOutputFolder)
)
}
\arguments{
\item{baseOutputFolder}{The base output folder of results}

\item{dbIds}{A vector of database IDs}

\item{packageName}{Name of package}

\item{finalOtputDir}{Final directory where the combined results should be saved}
}
\description{
Combines results across databases.
}
\details{
Combines results across databases.
}
